Split test_gui, rename integration to system
