+++ tests/system/gui/test_components.py
